order-service-python/requirement.txt
grpcio==1.44.0
grpcio-tools==1.44.0
protobuf==3.20.0
python-dateutil==2.8.2